node types: Preserve all supertypes in field type lists
